Code cleanup. Run script once per page load
Moved everything into a background page because all the js files need to
know about each other. bootstrap.js is a content script that fires a
message to the background page in window.onload to notify it of the page
load event, the current URL and the html of the page.